Reorganised tests for protocol and pacing systems.
